Move analysis to separate file
